Updated scoping of endpoints
