+++ ParraCore/Parra+Push.swift
@@ -1,0 +1,16 @@
+//
+//  Parra+Push.swift
+//  ParraCore
+//
+//  Created by Mick MacCallum on 6/24/23.
+//  Copyright © 2023 Parra, Inc. All rights reserved.
+//
+
+import Foundation
+
+public extension Parra {
+//    @MainActor
+//    class func initialize(config: ParraConfiguration = .default,
+//                          authProvider: ParraAuthenticationProviderType) {
+//    }
+}
